drawing characters instead of cmap now

+++ src/tact/Coordinate.cpp
@@ -43,12 +43,12 @@ int Coordinate::get_y() const {
 
 // // gets map coordiante so char can walk
 float Coordinate::get_map_x() const {
-     return (this->x)*32.0f;
+     return static_cast<float>(this->x)*32.0f;
  }
 
 // //gets map coordinate so char can walk
 float Coordinate::get_map_y() const {
-     return (this->y)*32.0f;
+     return static_cast<float>(this->y)*32.0f;
 }
